Move feincms_page_tags up, and stop reimplementing stuff in django-mptt
